tests: apply audit findings P1-F08 + P2-F07/F09/F11/F20 housekeeping
- Sync BUNDLED_SKILLS to 125 entries: remove stale open-pr/open-research-pr,
  add 15 missing planner/compose/prepare/reload entries, relocate generate-report
- Delete dead clone_isolation_repo fixture (tests/workspace/conftest.py)
- Merge tests/recipe/test__api.py (2 fns) into test_api.py; delete source
- Move tests/test_python_no_hardcoded_temp.py → tests/arch/ (fix _src_root depth)
- Delete tests/infra/test_recipe_write_advisor.py (subsumed by hooks/ version)
- Update tests/CLAUDE.md to reflect file moves/deletions
