+++ tests/workspace/test_skills.py
@@ -46,6 +46,8 @@ BUNDLED_SKILLS = [
     "bundle-local-report",
     "close-kitchen",
     "collapse-issues",
+    "compose-pr",
+    "compose-research-pr",
     "design-guards",
     "diagnose-ci",
     "dry-walkthrough",
@@ -69,6 +71,7 @@ BUNDLED_SKILLS = [
     "exp-lens-unit-interference",
     "exp-lens-validity-threats",
     "exp-lens-variance-stability",
+    "generate-report",
     "implement-experiment",
     "implement-worktree",
     "implement-worktree-no-merge",
@@ -85,15 +88,26 @@ BUNDLED_SKILLS = [
     "migrate-recipes",
     "open-integration-pr",
     "open-kitchen",
-    "open-pr",
-    "open-research-pr",
     "pipeline-summary",
     "plan-experiment",
     "plan-visualization",
+    "planner-analyze",
     "planner-elaborate-assignments",
+    "planner-elaborate-phase",
+    "planner-elaborate-wps",
+    "planner-extract-domain",
+    "planner-generate-phases",
+    "planner-reconcile-deps",
+    "planner-refine",
+    "planner-refine-assignments",
+    "planner-refine-phases",
+    "planner-refine-wps",
     "prepare-issue",
+    "prepare-pr",
+    "prepare-research-pr",
     "process-issues",
     "rectify",
+    "reload-session",
     "report-bug",
     "resolve-claims-review",
     "resolve-design-review",
@@ -129,7 +143,6 @@ BUNDLED_SKILLS = [
     "vis-lens-temporal",
     "vis-lens-uncertainty",
     "write-recipe",
-    "generate-report",
 ]
 
 # Internal-only skill documents: injected programmatically, never invocable as slash commands.
